chore: adjusted types and removed old type imports
