+++ apps/cranio-provider/src/types/app.ts
@@ -1,27 +1,27 @@
-import type { ICranioSchemas } from ".";
-import type { IDashboardPages, ICharts } from "./schema";
-import type { IOrganisations } from "./ErnDashboard";
+// import type { ICranioSchemas } from ".";
+// import type { IDashboardPages, ICharts } from "./schema";
+// import type { IOrganisations } from "./ErnDashboard";
 
-export interface IAppPage {
-  organisation: IOrganisations;
-  schemaNames: ICranioSchemas;
-  api: {
-    graphql: {
-      current: string;
-      public: string;
-      providers: string;
-    };
-  };
-}
+// export interface IAppPage {
+//   organisation: IOrganisations;
+//   schemaNames: ICranioSchemas;
+//   api: {
+//     graphql: {
+//       current: string;
+//       public: string;
+//       providers: string;
+//     };
+//   };
+// }
 
-export interface IOrganisationsResponse {
-  Organisations: IOrganisations[];
-}
+// export interface IOrganisationsResponse {
+//   Organisations: IOrganisations[];
+// }
 
-export interface IDashboardPagesResponse {
-  DashboardPages: IDashboardPages[];
-}
+// export interface IDashboardPagesResponse {
+//   DashboardPages: IDashboardPages[];
+// }
 
-export interface IChartsResponse {
-  Charts: ICharts[];
-}
+// export interface IChartsResponse {
+//   Charts: ICharts[];
+// }
