+++ apps/cranio-provider/src/views/cleft_lip_palate/level_1.vue
@@ -7,8 +7,7 @@ import type {
   ICharts,
   IChartData,
 } from "../../../../metadata-utils/src/viz/UiDashboard.js";
-import type { IAppPage } from "../../types/app";
-import type { ISiteErnCleftTypeCounts } from "../../types/index";
+import type { IAppPage, ISiteErnCleftTypeCounts } from "../../types";
 
 const props = defineProps<IAppPage>();
 const patientsByCleftType = ref<ISiteErnCleftTypeCounts>({
