+++ apps/cranio-provider/src/views/cleft_lip_palate/your_center.vue
@@ -21,8 +21,7 @@ import type {
   ICharts,
   IChartData,
 } from "../../../../metadata-utils/src/viz/UiDashboard";
-import type { IAppPage } from "../../types/app";
-import type { IKeyValuePair } from "../../types/index";
+import type { IAppPage, IKeyValuePair } from "../../types";
 
 const props = defineProps<IAppPage>();
 const loading = ref<boolean>(true);
