+++ apps/cranio-provider/src/views/craniosynostosis/all_centers/surgical.vue
@@ -25,8 +25,11 @@ import type {
   ICharts,
   IChartData,
 } from "../../../../../metadata-utils/src/viz/UiDashboard";
-import type { IAppPage } from "../../../types/app";
-import type { IKeyValuePair, IValueLabel } from "../../../types/index";
+import type {
+  IAppPage,
+  IKeyValuePair,
+  IValueLabel,
+} from "../../../types/index";
 
 const props = defineProps<IAppPage>();
 const loading = ref<boolean>(true);
